+++ test/protractor/spec/lforms_testpage.po.js
@@ -192,6 +192,7 @@ var TestPage = function() {
       q7: element(by.id('/q7/1')),
       q8: element(by.id('/q8/1')),
       q9: element(by.id('/q9/1')),
+      q99: element(by.id('/q99/1')),
 
       multiAnswers: element.all(by.css('.autocomp_selected li')),